Initial commit of stage1/2/3 and early services
Much of this is based on the output of s6-linux-init-maker but
with tool and stylistic changes. For example, I prefer busybox to the
s6-{linux,portable}-utils packages. The use of busybox, however, is
there mostly to make sure I do not use any GNU-specific features or
flags. Ideally any POSIX utility suite would work; I limited myself
where possible to flags specified in POSIX-2013.

Note that busybox kill is not conforming, so its use (in rc/shutdown)
is not prefixed. This, of course, is only helpful when /bin/kill is not
a link to busybox. A busybox-only implementation would use killall5.
Also, the signal-to-action mapping here does not exactly match what is
used by busybox halt/poweroff/reboot.

A getty on tty1 is set up to always run, so even if some early oneshot
in s6-rc fails, I can still log in and fix things. Other gettys will
be started after ok-init