Switch to a kill syntax that actually works
The original syntax looks better, is more explicit, and is specified by
POSIX, so it *should* work, but it doesn't. Procps kill interpretes the
-1 as an option (even after the "--" delimiter!) and prints a usage
message. If you give it UINT_MAX as a decimal instead, it happens to
work. Busybox, on the other hand, does not accept UINT_MAX on the
command line as a valid number, and claims to not support the full
syntax. Interestingly enough, it does (found out the hard way).

In any case, both implementations work correctly with -SIG instead of
-s SIG, so use that.