Add gettys (2-6)
These are part of a new bundle (ok-local), which represents the point at
which all "system"/infrastructure services are started, and a user can
log in.

These gettys additionally depend on ok-init for showing the correct
hostname from /etc/issue.

I would like the getty to clear the screen on logout, which busybox does
not do, so I might switch to util-linux agetty.