Expand update script for preprocessing (instantiation)
This causes services with an 'instances' file in their service dir to
be copied to <service>@<instance> directories for each line in the
'instances' file. The string '%I' is replaced by the instance name
anywhere it appears in the service dir (including data/ and env/).

Furthermore, a bundle is created with the uninstantiated service name
(that is, the name of the source directory in the filesystem). It
contains all of the service instances. This allows adding a service to
a target/runlevel/milestone without needing to know what instances
exist. 99.9% of the time, all instances of a service belong in the same
runlevel anyway.

Although @ and %I follow the conventions of systemd, this is not the
only reason they were chosen. @ was chosen to be distinct from -, to
point out that the instantiated service dir does not actually exist on
the filesystem (i.e. it doesn't look like just another service dir).
%I is short, clear (unlike e.g. @@ or {}), and is unlikely to appear
for other purposes. Either could easily be changed.

More escapes could be added in the future, though most are obtainable
from the environment (e.g. hostname, service name). The only one I can
think of that would need to be provided by the preprocessor is s6-rc
database compilation time.

Since this is now a larger script with more complex flow control, I
chose to implement it in POSIX shell.